Fixed compile issues on 64-bit platforms
closes #6

+++ include/Array.h
@@ -103,7 +103,7 @@ namespace Lucene
 		
 		int32_t hashCode() const
 		{
-			return (int64_t)container.get();
+			return (int32_t)(int64_t)container.get();
 		}
 		
 		TYPE& operator[] (int32_t i) const
