+++ include/Collection.h
@@ -207,7 +207,7 @@ namespace Lucene
 		
 		virtual int32_t hashCode()
 		{
-			return (int64_t)container.get();
+			return (int32_t)(int64_t)container.get();
 		}
 		
 		void swap(this_type& other)
