+++ include/Map.h
@@ -79,7 +79,7 @@ namespace Lucene
 		
 		virtual int32_t hashCode()
 		{
-			return (int64_t)mapContainer.get();
+			return (int32_t)(int64_t)mapContainer.get();
 		}
 		
 		operator bool() const
