+++ src/core/util/LuceneObject.cpp
@@ -37,7 +37,7 @@ namespace Lucene
     
     int32_t LuceneObject::hashCode()
     {
-        return (int64_t)this;
+        return (int32_t)(int64_t)this;
     }
     
     bool LuceneObject::equals(LuceneObjectPtr other)
